# inoa-measurement/src/main/resources/openapi/inoa-measurement-api.yaml
openapi: 3.0.3
info:
  title: Inoa Measurement API definition
  version: ${project.version}
  description: Definitions for Inoa Measurement.
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: GrayC GmbH
    url: https://grayc.de
servers:
  - url: http://measurement-service:8080
security:
  - Keycloak: []
  - BearerAuth: []
paths:
...
  /things:
    get:
      tags:
        - things
      parameters:
        - $ref: '#/components/parameters/PageablePage'
        - $ref: '#/components/parameters/PageableSize'
        - $ref: '#/components/parameters/Sort'
        - name: filter
          description: 'Search in name. Supports wildcard *, is case insensitive.'
          in: query
          schema:
            type: string
            maxLength: 10
            example: dvh
      summary: 'Find things.'
      description: 'Returns all things ordered by name.'
      operationId: findThings
      responses:
        200:
          description: 'Page with thing types.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingPage'
        401:
          description: 'Unauthorized.'
    post:
      tags:
        - things
      summary: 'Create thing type.'
      description: TODO
      operationId: createThing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThingCreate'
      responses:
        201:
          description: 'Thing created.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thing'
        400:
          description: 'Malformed.'
        401:
          description: 'Unauthorized.'
  /things/{thing_id}:
    parameters:
      - $ref: '#/components/parameters/ThingId'
    get:
      tags:
        - things
      summary: 'Find Thing.'
      description: TODO
      operationId: findThing
      responses:
        200:
          description: 'Thing found.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingDetail'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing not found.'
    put:
      tags:
        - things
      summary: 'Update Thing.'
      description: TODO
      operationId: updateThing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThingUpdate'
      responses:
        200:
          description: 'Thing updated.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Thing'
        400:
          description: 'Malformed.'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing not found.'
    delete:
      tags:
        - things
      summary: 'Delete thing.'
      description: TODO
      operationId: deleteThing
      responses:
        204:
          description: 'Thing deleted.'
        401:
          description: 'Unauthorized.'
        404:
          description: 'Thing type not found.'
  /things/by-gateway-id/{gateway_id}:
    parameters:
      - $ref: '#/components/parameters/GatewayId'
    get:
      tags:
        - things
      parameters:
        - $ref: '#/components/parameters/PageablePage'
        - $ref: '#/components/parameters/PageableSize'
        - $ref: '#/components/parameters/Sort'
        - name: filter
          description: 'Search in name. Supports wildcard *, is case insensitive.'
          in: query
          schema:
            type: string
            maxLength: 10
            example: dvh
      summary: 'Find all things for a gateway.'
      description: 'Returns all things.'
      operationId: findThingsByGatewayId
      responses:
        200:
          description: 'Page with thing types.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThingPage'
        401:
          description: 'Unauthorized.'
  /things/sync-to-gateway/{gateway_id}:
    parameters:
      - $ref: '#/components/parameters/GatewayId'
    get:
      tags:
        - things
      summary: 'Sync configuration to gateway.'
      description: 'TODO'
      operationId: syncConfigToGateway
      responses:
        204:
          description: 'accepted sync request.'
        401:
          description: 'Unauthorized.'
  /things/sync-to-gateway-sequential/{gateway_id}:
    parameters:
      - $ref: '#/components/parameters/GatewayId'
    get:
      tags:
        - things
      summary: 'Sync configuration sequentially to gateway.'
      description: 'TODO'
      operationId: syncConfigToGatewaySequential
      responses:
        204:
          description: 'accepted sync request.'
        401:
          description: 'Unauthorized.'
  /things/download-for-gateway/{gateway_id}:
    parameters:
      - $ref: '#/components/parameters/GatewayId'
    get:
      tags:
        - things
      summary: 'Sync configuration to gateway.'
      description: 'TODO'
      operationId: downloadConfigToGateway
      responses:
        200:
          description: 'json with thing config.'
          content:
            application/json:
              schema:
                type: object
        401:
          description: 'Unauthorized.'
  /things/download-for-gateway-legacy/{gateway_id}:
    parameters:
      - $ref: '#/components/parameters/GatewayId'
    get:
      tags:
        - things
      summary: 'Sync configuration to gateway with OS version <0.4.0'
      description: 'TODO'
      operationId: downloadConfigToGatewayLegacy
      responses:
        200:
          description: 'json with thing config.'
          content:
            application/json:
              schema:
                type: object
        401:
          description: 'Unauthorized.'
components:
  schemas:
    Timestamp:
      description: 'Common timestamp for created/updated timestamps.'
      type: string
      format: date-time
      example: '2021-01-01T00:00:00'
    Id:
      description: 'Id as technical reference (never changes).'
      type: string
      format: uuid
...
    GatewayId:
      description: 'Id as technical reference (never changes).'
      type: string
...
    Name:
      description: 'Name.'
      type: string
...
    ThingPage:
      description: 'Page for thing types.'
      type: object
      properties:
        content:
          description: 'List of entries on page.'
          type: array
          items:
            $ref: '#/components/schemas/Thing'
        total_size:
          description: 'Total available entries.'
          type: integer
          format: int32
          example: 0
      required:
        - content
        - total_size
    Thing:
      description: 'Thing with fields.'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          $ref: '#/components/schemas/Name'
        gateway_id:
          $ref: '#/components/schemas/GatewayId'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        config:
          description: 'config'
          additionalProperties: { }
        created:
          $ref: '#/components/schemas/Timestamp'
        updated:
          $ref: '#/components/schemas/Timestamp'
      required:
        - id
        - name
        - created
        - updated
    ThingDetail:
      description: 'Thing with fields and details.'
      type: object
      properties:
        id:
          $ref: '#/components/schemas/Id'
        name:
          $ref: '#/components/schemas/Name'
        gateway_id:
          $ref: '#/components/schemas/GatewayId'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        config:
          description: 'config'
          additionalProperties: { }
        created:
          $ref: '#/components/schemas/Timestamp'
        updated:
          $ref: '#/components/schemas/Timestamp'
      required:
        - id
        - name
        - created
        - updated
    ThingCreate:
      description: 'Thing to create.'
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        gateway_id:
          $ref: '#/components/schemas/GatewayId'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        config:
          description: 'config'
          additionalProperties: { }
      required:
        - name
        - thing_type_reference
    ThingUpdate:
      description: 'Thing to update.'
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        gateway_id:
          $ref: '#/components/schemas/GatewayId'
        thing_type_reference:
          description: 'thing_type_reference'
          type: string
        config:
          description: 'config'
          additionalProperties: { }
      required:
        - name
        - thing_type_reference
...
  parameters:
    PageablePage:
      description: 'Page number for pagination.'
      name: page
      in: query
      schema:
        type: integer
        format: int32
        minimum: 0
      example: 0
    PageableSize:
      description: 'Page size for pagination.'
      name: size
      in: query
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100
      example: 50
    Sort:
      description: 'Sorting.'
      name: sort
      in: query
      schema:
        type: array
        items:
          type: string
          pattern: '[a-zA-Z_]{2,10}(,(asc|desc|ASC|DESC))?'
      example:
        - name,asc
        - name,desc
...
    ThingId:
      name: thing_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Id'
    GatewayId:
      name: gateway_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/GatewayId'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    Keycloak:
      type: openIdConnect
      openIdConnectUrl: http://keycloak:8080/realms/inoa/.well-known/openid-configuration